feat: refresh expense in action
